Make multipart max memory usage more reasonable
This should ideally better support a workflow where we don't buffer the
entire message somewhere before making it available to the client. It is
not uncommon for the service image to have read-only filesystem access
so buffering on disk is not an option. It would be preferable to be able
to stream the message.